pyproject.toml
[tool.poetry]
name = "test-auth"
version = "0.1.0"
description = "Аутентификация"
authors = ["Андрей Балакирев <andrey_balakirevv@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.7.2"
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
bcrypt = "^4.1.3"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
